+++ src/shared/lib/index.ts
@@ -1,0 +1,3 @@
+export { default as cn } from '@/shared/lib/cn';
+export { default as getBoundedPositions } from '@/shared/lib/getBoundedPositions';
+export type { ExcludeArray } from '@/shared/lib/util.d.ts';
