+++ src/widgets/weather/ui/summary.tsx
@@ -1,4 +1,4 @@
-import { cn } from '@/shared/lib/cn';
+import { cn } from '@/shared/lib';
 
 interface SummaryProps {
   name: string;
