+++ src/widgets/weather/ui/weather-card.tsx
@@ -1,4 +1,4 @@
-import { cn } from '@/shared/lib/cn';
+import { cn } from '@/shared/lib';
 
 import WEATHER_DATA from '../model/DUMMY';
